LS-2313 - id to string and text fixes

+++ listshop/src/commonMain/kotlin/com/listshop/bff/data/remote/PutMergeRequest.kt
@@ -54,7 +54,7 @@ data class MergeItem @OptIn(ExperimentalSerializationApi::class) constructor(
             val tag = MergeTag.create(modelItem.tag)
             return MergeItem(
                 tag = tag,
-                itemId = modelItem.externalId,
+                itemId = modelItem.externalId.toLong(),
                 added = modelItem.added,
                 removed = modelItem.removed,
                 updated = modelItem.updatedOn,
